Batch-205/src/requirements.txt
backports.shutil-get-terminal-size==1.0.0
decorator==4.1.2
Django==1.8
django-crispy-forms==1.7.0
django-extensions==1.9.8
django-registration-redux==1.9
enum34==1.1.6
ipython==5.5.0
ipython-genutils==0.2.0
MySQL-python==1.2.5
mysqlclient==1.3.12
pathlib2==2.3.0
pexpect==4.3.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
PyMySQL==0.7.11
python-dateutil==2.6.1
scandir==1.6
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
typing==3.6.2
wcwidth==0.1.7
